+++ comicbook/conf/dao/common/ReadDAOConfig.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass, field
+from pyspark.sql.types import StructType
+
+
+@dataclass(frozen=True)
+class ReadDAOConfig:
+    path: str = None
+    schema: StructType = None
