+++ comicbook/dao/stats/StatsDAO.py
@@ -4,21 +4,12 @@ from pyspark.sql import DataFrame
 
 from comicbook.conf.dao.stats.StatsDAOConfig import StatsDAOConfig
 from comicbook.dao.common.DAOABCBase import DaoABC
+from comicbook.dao.common.read.ReadCsvDAOABC import ReadCsvDAOABC
 
 
 @dataclass(frozen=True)
-class StatsDAO(DaoABC):
+class StatsDAO(ReadCsvDAOABC, DaoABC):
     _config: StatsDAOConfig = field(repr=False, default=None)
-
-    def load(self) -> DataFrame:
-        return (self._spark_service.spark.read
-                                    .format("csv")
-                                    .option("mode", "DROPMALFORMED")
-                                    .option("mergeSchema", "true")
-                                    .schema(self._config.schema)
-                                    .option("header", True)
-                                    .load(self._config.path)
-         )
 
     def write(self, df: DataFrame) -> DataFrame:
         raise NotImplementedError("write is not implemented for StatsDAO")
